Fix rendering of arrowheads and in Safari
arrowheads: the cause of the intermittent bug was that #arrowhead was not a unique selector. Currently we now just add
the defs to only the SVG using them; this will need improvement to make it a common resource.

Safari does not support .innerHTML (which powers selection.html) so we now use the innersvg polyfill. Innversvg requires
classes be surrounded by (single?) quotes, which Chrome's .innerHTML did not, so we had to add those. The good news is
that Chrome and Safari now seem to be behaving identically, so it's less likely that a browser-specific issue will come
up.